HBASE-1252 Make atomic increment perform a binary increment
git-svn-id: https://svn.apache.org/repos/asf/hadoop/hbase/trunk@752284 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/hbase/client/HTable.java
@@ -1481,7 +1481,7 @@ public class HTable {
   }
   
   public long incrementColumnValue(final byte [] row, final byte [] column,
-      final int amount) throws IOException {
+      final long amount) throws IOException {
     return connection.getRegionServerWithRetries(
         new ServerCallable<Long>(connection, tableName, row) {
           public Long call() throws IOException {
